refactor: move state validations into their own package
Move the state-level resource validations into their own package, with one file per resource family.

Authorization checks are now applied ahead of value validations, so unauthorized callers see a permission error first.
