+++ internal/backend/runtime/omni/validations/testdata/secrets-broken.yaml
@@ -1,0 +1,3 @@
+cluster
+  id: rRxLgAjUwkp17VYlg5XSbTfKMpV4nn-i5yj-ZmSsxfM=
+  secret: tad0uMK9YeZBiSaN8+pMRH53ftmYl3Sdjv9Yko1f8Dc=
